another try on bigint

+++ db/schema.rb
@@ -60,13 +60,13 @@ ActiveRecord::Schema.define(:version => 20120504203645) do
     t.string   "status"
     t.boolean  "stale"
     t.date     "update_requirement"
-    t.integer  "order_value_cents",  :default => 0
+    t.bigint  "order_value_cents",  :default => 0
   end
 
   create_table "opportunity_records", :force => true do |t|
     t.integer  "opportunity_id"
     t.string   "salesman"
-    t.integer  "order_value_cents", :default => 0
+    t.bigint  "order_value_cents", :default => 0
     t.date     "order_date"
     t.integer  "days_to_order"
     t.string   "status"
@@ -86,10 +86,10 @@ ActiveRecord::Schema.define(:version => 20120504203645) do
 
   create_table "targets", :force => true do |t|
     t.integer  "fiscal_year"
-    t.integer  "q1_cents",        :default => 0
-    t.integer  "q2_cents",        :default => 0
-    t.integer  "q3_cents",        :default => 0
-    t.integer  "q4_cents",        :default => 0
+    t.bigint  "q1_cents",        :default => 0
+    t.bigint  "q2_cents",        :default => 0
+    t.bigint  "q3_cents",        :default => 0
+    t.bigint  "q4_cents",        :default => 0
     t.string   "targetable_type"
     t.integer  "targetable_id"
     t.datetime "created_at"
